core: accept procedural core path and trim PR guardrails
