+++ .github/pull_request_template.md
@@ -2,29 +2,6 @@
 
 What does this PR do and why?
 
-## Bug class
-
-Does this PR fix or prevent a known recurring bug class?
-
-Reference the source of truth at [`.github/bug-classes.yml`](bug-classes.yml) and the human mirror at [`docs/engineering/bug-classes.md`](../docs/engineering/bug-classes.md).
-
-> Addressed bug class(es): ___________
-
-## Regression guardrail
-
-What permanent guardrail (contract test, CI gate, automated check) does this PR add or strengthen to prevent this bug class from recurring?
-
-> Regression guardrail: ___________
-
-## Duplicate / Recurring Issue Handling
-
-If this PR handles duplicate issues, a recurrence, or an umbrella issue, record the disposition.
-
-> Type: duplicate | recurrence | umbrella | new
-> Canonical issue: ___________
-> Related issues to close/update: ___________
-> Closing comment summary: ___________
-
 ## Scope
 
 - [ ] Bug fix
@@ -40,18 +17,10 @@ List the main files or modules changed (one per line):
 
 ## Validation / Checks Run
 
-What did you run to verify this change?
-
-- [ ] `make check`
-- [ ] `make test-unit`
-- [ ] `make test`
-- [ ] Focused pytest on touched files (`uv run pytest <path> -q`)
-- [ ] Other: ___________
-
 Checks run (paste the command and output summary):
 > ___________
 
-If any check above was skipped, state the reason explicitly:
+If any check was skipped, state the reason explicitly:
 > ___________
 
 ## Runtime Impact
