+++ src/core/__init__.py
@@ -1,18 +1,22 @@
 """Core application module exports."""
 
-from .assistant import (
+from .contracts import (
     AssistantError,
+    AssistantRequest,
     AssistantResult,
     CoreDependencies,
     CrmAction,
     UserContext,
-    run_assistant_request,
 )
 
 
 def __getattr__(name: str) -> object:
-    """Load legacy pipeline exports lazily to keep assistant imports lightweight."""
+    """Load runtime-bearing exports lazily to keep core contracts import-safe."""
 
+    if name == "run_assistant_request":
+        from .assistant import run_assistant_request
+
+        return run_assistant_request
     if name == "RAGPipeline":
         from .pipeline import RAGPipeline
 
@@ -22,6 +26,7 @@ def __getattr__(name: str) -> object:
 
 __all__ = [
     "AssistantError",
+    "AssistantRequest",
     "AssistantResult",
     "CoreDependencies",
     "CrmAction",
