+++ telegram_bot/services/cache_policy.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 from time import time
 from typing import Any
 
-from telegram_bot.services.grounding_policy import semantic_cache_safe_reuse_allowed
+from src.runtime.grounding.policy import semantic_cache_safe_reuse_allowed
 from telegram_bot.services.query_filter_signal import build_filter_signature
