+++ tests/data/known_runtime_telegram_bot_couplings.json
@@ -1,0 +1,19 @@
+{
+  "src/runtime/graph/builder.py": [
+    "telegram_bot.graph.graph:build_graph"
+  ],
+  "src/runtime/pipeline/assistant_pipeline.py": [
+    "telegram_bot.services.generate_response"
+  ],
+  "src/runtime/pipeline/rag.py": [
+    "telegram_bot.graph.config",
+    "telegram_bot.observability",
+    "telegram_bot.services.bge_m3_query_bundle",
+    "telegram_bot.services.cache_policy",
+    "telegram_bot.services.metrics",
+    "telegram_bot.services.query_filter_signal",
+    "telegram_bot.services.query_preprocessor",
+    "telegram_bot.services.rag_core",
+    "telegram_bot.services.small_to_big"
+  ]
+}
